+++ src/components/vis/utils.jsx
@@ -1,0 +1,31 @@
+import Works from "../../assets/works_v03.json";
+import Composer from "../../assets/composers_v02.json";
+
+export const matchedDataByIds = Works.map((work) => {
+  const composerInfo = Composer.find((c) => c.name === work.composer);
+
+  return {
+    composer: work.composer,
+    id: work.id,
+    title: work.title,
+    year: work.year,
+    lat: composerInfo ? composerInfo.latitude : null,
+    lon: composerInfo ? composerInfo.longitude : null,
+    nationality: composerInfo ? composerInfo.nationality : null,
+    name:
+      (composerInfo ? composerInfo.nationality : "Unknown") +
+      "/" +
+      work.year +
+      "/" +
+      work.composer +
+      "/" +
+      work.title,
+  };
+});
+
+export const getComposerFromId = (composerId) => {
+  const work = matchedDataByIds.find((item) => item.id === composerId);
+  return work ? work : { lat: null, lon: null };
+};
+
+
